# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MailUp - Backend Test",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Mostrar categorias existentes",
                "description": "Permite detallar los categorias registradas.",
                "operationId": "e225c2b7eb5daf7fb16e00f4f07ff030",
                "responses": {
                    "200": {
                        "description": "Operacion Exitosa."
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Registrar nueva Categoría",
                "description": "Permite registrar una nueva Categoría.",
                "operationId": "dd76b8d73b7ea8b4951f03d7c0904c92",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre de la Categoría",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "422": {
                        "description": "El campo ya esta en uso."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener información de una Categoría",
                "description": "Devuelve datos de una Categoría específica según Id. ",
                "operationId": "9c4934d1b68a6d4440ec72cfc8ae7074",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese Id de una Categoría existente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa."
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    },
                    "404": {
                        "description": "Recurso no encontrado."
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Actualizar una Categoría existente",
                "description": "Permite modificar datos de una Categoría existente.",
                "operationId": "988e8695bc991d7f8e40131db5ba7a76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de Categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre de la Categoría",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "422": {
                        "description": "El campo ya esta en uso."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Eliminar Categoría existente",
                "description": "Permite eliminar un registro y no devuelve contenido.",
                "operationId": "37196d259228d2274dd9dbef8b00e547",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de Categoría a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa."
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    },
                    "404": {
                        "description": "Recurso no encontrado."
                    }
                }
            }
        },
...
    }
}